cli: rename discovery.ts
